finish steps 1,2,3
