+++ helpers/index.js
@@ -1,5 +1,7 @@
-const HttpError = require("./HttpError");
+const HttpError = require("./httpError");
+const ctrlWrapper = require("./ctrl_wrapper");
 
 module.exports = {
   HttpError,
+  ctrlWrapper,
 };
