+++ models/contacts.json
@@ -6,12 +6,6 @@
     "phone": "(992) 914-3792"
   },
   {
-    "id": "drsAJ4SHPYqZeG-83QTVW",
-    "name": "Kennedy Lane",
-    "email": "mattis.Cras@nonenimMauris.net",
-    "phone": "(542) 451-7038"
-  },
-  {
     "id": "vza2RIzNGIwutCVCs4mCL",
     "name": "Wylie Pope",
     "email": "est@utquamvel.net",
@@ -52,5 +46,11 @@
     "name": "Alec Howard",
     "email": "Donec.elementum@scelerisquescelerisquedui.net",
     "phone": "(748) 206-2688"
+  },
+  {
+    "id": "B-XRq_hb3Oe_4LMyljZ4n",
+    "name": "Lewis Creator",
+    "email": "dui.in@egetl.ca",
+    "phone": "(294) 840-6622"
   }
 ]
