+++ routes/api/contacts.js
@@ -2,22 +2,7 @@ const express = require("express");
 const contacts = require("../../models/contacts");
 const router = express.Router();
 const { HttpError } = require("../../helpers");
-const Joi = require("joi");
-
-const addSchema = Joi.object({
-  name: Joi.string()
-    .min(2)
-    .required()
-    .messages({ "any.required": "missing required name field" }),
-  email: Joi.string()
-    .min(6)
-    .required()
-    .messages({ "any.required": "missing required email field" }),
-  phone: Joi.string()
-    .min(8)
-    .required()
-    .messages({ "any.required": "missing required phone field" }),
-});
+const addSchema = require("../../schemas/contacts");
 
 router.get("/", async (req, res, next) => {
   try {
